+++ src/global-realm.ts
@@ -1,0 +1,7 @@
+import { Realm } from './realm';
+
+const globalRealm = new Realm();
+
+const { define, manifest, persist } = globalRealm;
+
+export { define, manifest, persist };