+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from './errors';
+export * from './global-realm';
 export * from './realm';
 export * from './ref';
 export * from './types';