+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added function-based API using a global `Realm`
+
 ### Changed
 
 - Entity names are now enforced to be unique per `Realm`